fix(test): move sparkline dark-mode test to css-regression.spec.ts, strengthen assertion (#653)
- Remove `sparklines visible in dark mode` from visual-regression.spec.ts (wrong file for a DOM/CSS assertion test)
- Revert appUrl import in visual-regression.spec.ts (was only added to support the misplaced test)
- Add the test to css-regression.spec.ts using the adminTest fixture pattern
- Strengthen assertion from `not.toBe('rgb(0, 0, 0)')` to `.toMatch(/^rgb/)` so stroke:'none' is also caught
